# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/session": {
            "get": {
                "description": "Get list of active sessions with cookies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "List sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.ResponseSession"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new session and save cookies to MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Login session",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete session from MongoDB and close browser",
                "tags": [
                    "session"
                ],
                "summary": "Delete session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/session/check": {
            "get": {
                "description": "Check if session is still valid",
                "tags": [
                    "session"
                ],
                "summary": "Check session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/api/session/click": {
            "post": {
                "description": "Click an element in the session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Click element",
                "parameters": [
                    {
                        "description": "Click request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ClickRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clicked",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ClickRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "selector": {
                    "type": "string"
                }
            }
        },
        "controller.Request": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "passs": {
                    "type": "string"
                }
            }
        },
        "controller.ResponseSession": {
            "type": "object",
            "properties": {
                "cookies": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "screenshot": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}